FileAPI/filereaderSync_readAsText.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>FileAPI Test: filereaderSync_readAsText</title>
<link rel="author" title="Intel" href="http://www.intel.com">
<link rel="help" href="http://www.w3.org/TR/FileAPI/#dfn-readAsTextSync">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>

<script>
async_test(function(t) {
  var worker = new Worker("filereaderSync.js");
  worker.postMessage("fileReaderSync_readAsText");
  worker.onmessage = t.step_func(function (event) {
    assert_equals(event.data, "test");
    t.done();
  });
}, "Check read a blob text using readAsText by FileReaderSync interface");
</script>
